Implement retrying dispatcher requests
If the first request fails it will retry
after a minute, hour, one day. If it continues
to fail after one day it will mark the server
as not reachable in the database.

As soon as the localhost receives a valid
message again it will mark the server as
reachable and the cycle starts anew